rebased month name displayer on year view month-containers
