--- src/Components/year-view.tsx
@@ -1,9 +1,0 @@
-import MonthContainers from "@/Components/YearViewComponents/month-containers.tsx";
-
-export default function YearView() {
-	return (
-		<div className=" p-4 rounded-b-2xl border  flex  justify-center ">
-			<MonthContainers />
-		</div>
-	);
-}
