+++ src/components/year-view-components/month-name-displayer.tsx
@@ -1,0 +1,25 @@
+import { useCalendar } from "@/context/calendar-context.tsx";
+import { formatDate } from "date-fns";
+
+export default function monthNameDisplayer({ month }: { month: Date }) {
+  const formattedMonth = formatDate(month, "MMMM");
+
+  const { setDate, setView } = useCalendar();
+
+  const handleMonthClick = (month: Date) => {
+    setView("month");
+    setDate(month);
+  };
+
+  return (
+    <button
+      type="button"
+      onClick={() => {
+        handleMonthClick(month);
+      }}
+      className="text-[16px] font-semibold text-gray-800 dark:text-gray-200  text-center px-20 py-2 border cursor-pointer  rounded-t-2xl hover:bg-gray-100"
+    >
+      {formattedMonth}
+    </button>
+  );
+}
